[cont-free] feat: separar inventario en módulo independiente con sidebar
