+++ sysgd-cont/src/app/app.component.css
@@ -131,6 +131,38 @@ h1 {
   margin: 30px auto;
 }
 
+
+.app-layout {
+  display: grid;
+  grid-template-columns: 260px minmax(0, 1fr);
+  gap: 12px;
+  align-items: start;
+}
+
+.sidebar {
+  position: sticky;
+  top: 12px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.sidebar button {
+  background: #ffffff;
+  color: #111827;
+  border: 1px solid #d1d5db;
+  text-align: left;
+}
+
+.sidebar button.active {
+  background: #2563eb;
+  color: #ffffff;
+  border-color: #2563eb;
+}
+
+.content-area {
+  min-width: 0;
+}
 .tabs {
   display: flex;
   gap: 8px;
@@ -328,6 +360,16 @@ button {
   border-radius: 8px;
   padding: 10px;
   font-size: 13px;
+}
+
+@media (max-width: 980px) {
+  .app-layout {
+    grid-template-columns: 1fr;
+  }
+
+  .sidebar {
+    position: static;
+  }
 }
 
 @media (max-width: 840px) {
