+++ sysgd-cont/src/app/app.component.html
@@ -25,80 +25,91 @@
   />
 
   <ng-container *ngIf="sessionReady && currentUser">
-    <nav class="tabs">
-      <button [class.active]="activeTab === 'generales'" (click)="activeTab = 'generales'">1. Generales</button>
-      <button [class.active]="activeTab === 'movimientos'" (click)="activeTab = 'movimientos'">2-3. Ingresos/Gastos</button>
-      <button [class.active]="activeTab === 'tributos'" (click)="activeTab = 'tributos'">4. Tributos</button>
-      <button [class.active]="activeTab === 'resumen'" (click)="activeTab = 'resumen'">Resumen DJ</button>
-      <button [class.active]="activeTab === 'recursos'" (click)="activeTab = 'recursos'">Recursos</button>
-    </nav>
+    <section class="app-layout">
+      <aside class="sidebar card">
+        <button [class.active]="activeTab === 'generales'" (click)="activeTab = 'generales'">1. Generales</button>
+        <button [class.active]="activeTab === 'movimientos'" (click)="activeTab = 'movimientos'">2-3. Ingresos/Gastos</button>
+        <button [class.active]="activeTab === 'inventario'" (click)="activeTab = 'inventario'">Módulo Compras/Ventas</button>
+        <button [class.active]="activeTab === 'tributos'" (click)="activeTab = 'tributos'">4. Tributos</button>
+        <button [class.active]="activeTab === 'resumen'" (click)="activeTab = 'resumen'">Resumen DJ</button>
+        <button [class.active]="activeTab === 'recursos'" (click)="activeTab = 'recursos'">Recursos</button>
+      </aside>
 
-    <section class="card instructions-card">
-      <details>
-        <summary>Instrucciones de llenado y conservación del registro</summary>
-        <p class="subtitle">Objetivo: facilitar el registro de operaciones para la Declaración Jurada. Todos los importes se registran en CUP.</p>
-        <ul>
-          <li>El registro debe conservarse limpio, en buen estado y actualizado.</li>
-          <li>Puede llevarse en formato digital y debe conservarse por 5 años junto con sus comprobantes.</li>
-          <li>En ingresos/gastos se anota el día del mes y el importe del día correspondiente.</li>
-          <li>Si cobras o pagas en divisa, convierte a CUP según la tasa vigente del BCC.</li>
-          <li>No se permiten borrones: si hay error, usa el método E/ (error) y C/ (corrección).</li>
-          <li>Al cierre de cada mes, anula filas no usadas y totaliza el mes.</li>
-          <li>En tributos, registra mensualmente cada concepto pagado y su total por mes.</li>
-          <li>En la DJ, estos importes se deducen en las secciones y casillas correspondientes.</li>
-        </ul>
-      </details>
+      <div class="content-area">
+        <section class="card instructions-card">
+          <details>
+            <summary>Instrucciones de llenado y conservación del registro</summary>
+            <p class="subtitle">Objetivo: facilitar el registro de operaciones para la Declaración Jurada. Todos los importes se registran en CUP.</p>
+            <ul>
+              <li>El registro debe conservarse limpio, en buen estado y actualizado.</li>
+              <li>Puede llevarse en formato digital y debe conservarse por 5 años junto con sus comprobantes.</li>
+              <li>En ingresos/gastos se anota el día del mes y el importe del día correspondiente.</li>
+              <li>Si cobras o pagas en divisa, convierte a CUP según la tasa vigente del BCC.</li>
+              <li>No se permiten borrones: si hay error, usa el método E/ (error) y C/ (corrección).</li>
+              <li>Al cierre de cada mes, anula filas no usadas y totaliza el mes.</li>
+              <li>En tributos, registra mensualmente cada concepto pagado y su total por mes.</li>
+              <li>En la DJ, estos importes se deducen en las secciones y casillas correspondientes.</li>
+            </ul>
+          </details>
+        </section>
+
+        <app-generales-section
+          *ngIf="activeTab === 'generales'"
+          [generalesForm]="generalesForm"
+          [provinces]="provinces"
+          [fiscalMunicipios]="fiscalMunicipios"
+          [legalMunicipios]="legalMunicipios"
+          (saveGenerales)="saveGenerales()"
+        />
+
+        <app-movimientos-section
+          *ngIf="activeTab === 'movimientos'"
+          [movForm]="movForm"
+          [months]="months"
+          [selectedMonth]="selectedMonth"
+          [registro]="registro"
+          [selectedMonthIngresosTotal]="selectedMonthIngresosTotal"
+          [selectedMonthGastosTotal]="selectedMonthGastosTotal"
+          [trackByDayAmount]="trackByDayAmount"
+          (saveMovement)="saveMovement()"
+        />
+
+        <app-inventario-section
+          *ngIf="activeTab === 'inventario'"
+          [months]="months"
+          [selectedMonth]="selectedMonth"
+          [registro]="registro"
+          [historialOperacionesMes]="historialOperacionesMes"
+          (selectedMonthChange)="selectedMonth = $event"
+          (addProductoInventarioEvent)="addProductoInventario($event)"
+          (registrarOperacionInventarioEvent)="registrarOperacionInventario($event)"
+        />
+
+        <app-tributos-section
+          *ngIf="activeTab === 'tributos'"
+          [tributoForm]="tributoForm"
+          [months]="months"
+          [tributosRows]="registro.tributos"
+          (autoCalculate)="autoCalculateTributos()"
+          (saveTributos)="saveTributos()"
+        />
+
+        <app-resumen-section
+          *ngIf="activeTab === 'resumen'"
+          [report]="report"
+          [threshold]="threshold"
+          [alerts]="alerts"
+          [djPreview]="djPreview"
+          [pdfLoading]="pdfLoading"
+          [backupMessage]="backupMessage"
+          (downloadPdf)="downloadPdf()"
+          (exportJson)="exportRegistroJson()"
+          (importJson)="importRegistroJson($event)"
+        />
+
+        <app-resources-section *ngIf="activeTab === 'recursos'" />
+      </div>
     </section>
-
-    <app-generales-section
-      *ngIf="activeTab === 'generales'"
-      [generalesForm]="generalesForm"
-      [provinces]="provinces"
-      [fiscalMunicipios]="fiscalMunicipios"
-      [legalMunicipios]="legalMunicipios"
-      (saveGenerales)="saveGenerales()"
-    />
-
-    <app-movimientos-section
-      *ngIf="activeTab === 'movimientos'"
-      [movForm]="movForm"
-      [months]="months"
-      [selectedMonth]="selectedMonth"
-      [registro]="registro"
-      [selectedMonthIngresosTotal]="selectedMonthIngresosTotal"
-      [selectedMonthGastosTotal]="selectedMonthGastosTotal"
-      [historialOperacionesMes]="historialOperacionesMes"
-      [trackByDayAmount]="trackByDayAmount"
-      (saveMovement)="saveMovement()"
-      (addProductoInventarioEvent)="addProductoInventario($event)"
-      (registrarOperacionInventarioEvent)="registrarOperacionInventario($event)"
-    />
-
-    <app-tributos-section
-      *ngIf="activeTab === 'tributos'"
-      [tributoForm]="tributoForm"
-      [months]="months"
-      [tributosRows]="registro.tributos"
-      (autoCalculate)="autoCalculateTributos()"
-      (saveTributos)="saveTributos()"
-    />
-
-    <app-resumen-section
-      *ngIf="activeTab === 'resumen'"
-      [report]="report"
-      [threshold]="threshold"
-      [alerts]="alerts"
-      [djPreview]="djPreview"
-      [pdfLoading]="pdfLoading"
-      [backupMessage]="backupMessage"
-      (downloadPdf)="downloadPdf()"
-      (exportJson)="exportRegistroJson()"
-      (importJson)="importRegistroJson($event)"
-    />
-
-    <app-resources-section
-      *ngIf="activeTab === 'recursos'"
-    />
   </ng-container>
 
   <section class="promo-banner promo-banner-bottom" *ngIf="showPromoBanner">
