+++ sysgd-cont/src/app/features/inventario-section/inventario-section.component.html
@@ -1,0 +1,72 @@
+<section class="card">
+  <div class="inventario-header">
+    <h2>Módulo de Compras y Ventas</h2>
+    <label>Mes
+      <select [ngModel]="selectedMonth" (ngModelChange)="selectedMonthChange.emit($event)">
+        <option *ngFor="let m of months" [ngValue]="m">{{ m }}</option>
+      </select>
+    </label>
+  </div>
+
+  <div class="inventario-tabs">
+    <button type="button" [class.active]="inventarioTab === 'venta'" (click)="inventarioTab='venta'">Venta</button>
+    <button type="button" [class.active]="inventarioTab === 'compra'" (click)="inventarioTab='compra'">Compra</button>
+    <button type="button" [class.active]="inventarioTab === 'historial'" (click)="inventarioTab='historial'">Historial</button>
+  </div>
+
+  <div *ngIf="inventarioTab !== 'historial'" class="form-grid">
+    <label>Tipo producto
+      <select [(ngModel)]="productoTipo">
+        <option value="venta">Venta</option>
+        <option value="compra">Compra</option>
+      </select>
+    </label>
+    <label>Nombre <input [(ngModel)]="productoNombre" /></label>
+    <label>Precio CUP <input type="number" min="0.01" step="0.01" [(ngModel)]="productoPrecio" /></label>
+    <label>Unidad <input [(ngModel)]="productoUnidad" /></label>
+    <div class="actions full">
+      <button type="button" (click)="addProductoInventario()">Agregar producto</button>
+    </div>
+
+    <label>Tipo operación
+      <select [(ngModel)]="operacionTipo" (change)="operacionProductoId=''">
+        <option value="venta">Venta</option>
+        <option value="compra">Compra</option>
+      </select>
+    </label>
+    <label>Producto
+      <select [(ngModel)]="operacionProductoId">
+        <option value="">Selecciona...</option>
+        <option *ngFor="let p of productosOperacion" [value]="p.id">{{ p.nombre }} ({{ p.precio | number:'1.2-2' }} / {{ p.unidad }})</option>
+      </select>
+    </label>
+    <label>Cantidad <input type="number" min="0.01" step="0.01" [(ngModel)]="operacionCantidad" /></label>
+    <label>Fecha <input type="date" [(ngModel)]="operacionFecha" /></label>
+    <div class="actions full">
+      <button type="button" (click)="registrarOperacionInventario()">Registrar operación</button>
+    </div>
+  </div>
+
+  <div class="table-section" *ngIf="inventarioTab === 'historial'">
+    <h3>Historial {{ selectedMonth }}</h3>
+    <div class="table-wrapper">
+      <table class="mov-table">
+        <thead>
+          <tr><th>Fecha</th><th>Tipo</th><th>Producto</th><th>Cantidad</th><th>Total</th></tr>
+        </thead>
+        <tbody>
+          <tr *ngFor="let op of historialOperacionesMes">
+            <td>{{ op.fecha }}</td>
+            <td>{{ op.tipo }}</td>
+            <td>{{ op.nombreProducto }}</td>
+            <td>{{ op.cantidad | number:'1.0-2' }} {{ op.unidad }}</td>
+            <td>{{ op.total | currency:'CUP':'symbol':'1.2-2' }}</td>
+          </tr>
+          <tr *ngIf="historialOperacionesMes.length === 0">
+            <td colspan="5" class="empty-cell">Sin operaciones en este mes.</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </div>
+</section>
